Separacao de implementacao e declaracao
